mydjango23/requirements.txt
django~=3.2.0
pillow
django-debug-toolbar>=3.2.3
django-bootstrap-v5
tablib[xlsx]